requirements.txt
pdfplumber
jwt
numpy
vertexai
transformers
requests
pillow
aiohttp
ipython
librosa
chainlit
soundfile
pydub
ffmpeg
torch
google-generativeai
kenlm
pyctcdecode
config
bs4
markdown
latex2mathml
llama_index
llama_index.llms.anthropic
pinecone
llama-index-vector-stores-pinecone
llama_index.llms.huggingface
google-generativeai
llama-index>=0.9.31
pinecone-client>=3.0.0
ipython
llama-index-embeddings-cohere
llama-index-embeddings-huggingface
llama-index-embeddings-langchain
llama-index
llama-index-llms-anthropic
llama-index-multi-modal-llms-anthropic
docx2txt
langchain-community
functions-framework
ffprobe
